binary search refactoring

+++ src/main/java/com/svetanis/algorithms/search/binary/frequency/FirstAndLastOccurrence34.java
@@ -8,15 +8,19 @@ public final class FirstAndLastOccurrence34 {
 	// Time Complexity: O(log n)
 
 	public static int[] firstAndLast(int[] a, int target) {
+		// long: target + 1 is the whole trick, and it wraps to
+		// MIN_VALUE when target is Integer.MAX_VALUE
 		int left = firstOccurrence(a, target);
-		int right = firstOccurrence(a, target + 1);
+		int right = firstOccurrence(a, target + 1L);
 		if (left == right) {
 			return new int[] { -1, -1 };
 		}
 		return new int[] { left, right - 1 };
 	}
 
-	private static int firstOccurrence(int[] a, int target) {
+	// returns how many elements are strictly below target, which is
+	// also the index the first one would sit at
+	private static int firstOccurrence(int[] a, long target) {
 		int left = 0;
 		int right = a.length - 1;
 		while (left <= right) {
